KashY_IP_Rep_Checker/requirements.txt
requests
geoip2
python-dotenv